[ADD] frontend date and number formatting #13 and other fixes

+++ backend/src/tenant/inventory/model.rs
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-
+use bigdecimal::BigDecimal;
 use chrono::{DateTime, Local};
 use serde::{Deserialize, Serialize};
 use sqlx::FromRow;
@@ -27,11 +27,11 @@ pub struct Inventory {
     pub id: Uuid,
     pub product_id: Uuid,
     pub warehouse_id: Uuid,
-    pub quantity_on_hand: i32,
-    pub quantity_reserved: i32,
-    pub quantity_available: i32,
-    pub minimum_stock: Option<i32>,
-    pub maximum_stock: Option<i32>,
+    pub quantity_on_hand: BigDecimal,
+    pub quantity_reserved: BigDecimal,
+    pub quantity_available: BigDecimal,
+    pub minimum_stock: Option<BigDecimal>,
+    pub maximum_stock: Option<BigDecimal>,
     pub currency_code: String,
     pub status: String,
     pub created_by_id: Uuid,
@@ -47,11 +47,11 @@ pub struct InventoryResolved {
     pub product: String,
     pub warehouse_id: Uuid,
     pub warehouse: String,
-    pub quantity_on_hand: i32,
-    pub quantity_reserved: i32,
-    pub quantity_available: i32,
-    pub minimum_stock: Option<i32>,
-    pub maximum_stock: Option<i32>,
+    pub quantity_on_hand: BigDecimal,
+    pub quantity_reserved: BigDecimal,
+    pub quantity_available: BigDecimal,
+    pub minimum_stock: Option<BigDecimal>,
+    pub maximum_stock: Option<BigDecimal>,
     pub currency_code: String,
     pub currency: String,
     pub status: String,
